# Implement custom fold for `WhileSome`