--- a/src/adaptors/mod.rs
+++ b/src/adaptors/mod.rs
@@ -581,6 +581,17 @@ where
     fn size_hint(&self) -> (usize, Option<usize>) {
         (0, self.iter.size_hint().1)
     }
+
+    fn fold<B, F>(self, acc: B, f: F) -> B
+    where
+        Self: Sized,
+        F: FnMut(B, Self::Item) -> B,
+    {
+        self.iter
+            .take_while(|opt| opt.is_some())
+            .map(|item| item.unwrap())
+            .fold(acc, f)
+    }
 }
 
 /// An iterator to iterate through all combinations in a `Clone`-able iterator that produces tuples