#755 

This PR introduces a custom fold method for `WhileSome`. The optimization hinges on the underlying iterator's `fold`:

With Custom `fold`: If the underlying iterator provides its own specialized `fold`, the performance gain depends on its implementation.
Without Custom `fold`: In the absence of a specialized fold for the underlying iterator, the performance remains on par with the default.

Benchmark Results:

Default fold: time: [497.09 ns 497.45 ns 497.82 ns]
Custom fold: time: [456.09 ns 458.15 ns 460.80 ns]