@Philippe-Cholet 2023-10-11
I think the bench part should be

    let result = data.fold(String::new(), |mut acc, ch| {
        acc.push(ch);
        acc
    });
    assert_eq!(result, "0123456789abcdef"); // or result.as_str() if needed