@Philippe-Cholet 2023-10-11
Each time in `fold`, `to_string` here creates a non-empty string and therefore heap-allocates (slow) while `push(ch)` merely mutates the existing `acc` string, so yes way heavier and the reason behind my quick comment, I should have said something.
The `assert_eq` was probably a bit slow too for allocating to a vector while my suggestion would not.